test(prompt-area): hoist jsdom Range polyfill to global setup, share caret helpers
The Range.getBoundingClientRect polyfill was copy-pasted into five test
files; move it to vitest.setup.ts since it's a jsdom environment gap, not a
per-file concern. Extract the duplicated placeCursorAtEnd/placeCursor
helpers into a shared test-helpers module. No behavior change.
